+++ LCS/delta/LCS.h
@@ -37,5 +37,5 @@ protected:
   typedef map<string, INDEXES> STRING2INDEXES;
   typedef deque<INDEXES*> MATCHES;
 
-  uint64_t Pairs(MATCHES& matches, shared_ptr<Pair> *pairs);
+  uint32_t Pairs(MATCHES& matches, shared_ptr<Pair> *pairs);
 };
